Update SearchResults.js
Submitting on behalf of Henry. Created SearchResults page with a grid containing top results on left, and most related on the right. Interactive Search bar that as you type left side of grid shrinks to more accurate results. "View Details" hyperlink in each anime result that leads to a single page containing all of the Anime title's information and rating.